restructured toplevel directories
