--- experiments/results/1_sterilised_problem/results.csv
@@ -1,73 +1,0 @@
-method;d;nu;jit;time_attempt_step
-ek1_reference;4;8;False;0.07339096010000504
-ek1_reference;4;8;True;0.00019989500001429405
-ek1_reference;8;8;False;0.07944169200000033
-ek1_reference;8;8;True;0.000888162900014322
-ek1_reference;16;8;False;0.17077787630000785
-ek1_reference;16;8;True;0.010903117800012296
-ek1_reference;32;8;False;0.20356393320000735
-ek1_reference;32;8;True;0.0352766875000043
-ek1_reference;64;8;False;0.265103093599987
-ek1_reference;64;8;True;0.14365970420001303
-ek1_reference;128;8;False;0.9478633799999898
-ek1_reference;128;8;True;0.8231581518999974
-ek1_diagonal;4;8;False;0.029496564099986244
-ek1_diagonal;4;8;True;7.561949998944329e-05
-ek1_diagonal;8;8;False;0.024705627699995604
-ek1_diagonal;8;8;True;9.84965000043303e-05
-ek1_diagonal;16;8;False;0.03632258600000569
-ek1_diagonal;16;8;True;0.00014026179999291344
-ek1_diagonal;32;8;False;0.038862749800000526
-ek1_diagonal;32;8;True;0.00022568920001049265
-ek1_diagonal;64;8;False;0.03871931179999137
-ek1_diagonal;64;8;True;0.0003977041999860376
-ek1_diagonal;128;8;False;0.0408961354999974
-ek1_diagonal;128;8;True;0.0008911421999982849
-ek1_diagonal;256;8;False;0.06090621749999627
-ek1_diagonal;256;8;True;0.002012080999998034
-ek1_diagonal;512;8;False;0.1081998773000123
-ek1_diagonal;512;8;True;0.0053514774999939615
-ek1_diagonal;1024;8;False;0.24231756550000227
-ek1_diagonal;1024;8;True;0.015687176499977796
-ek1_truncated;4;8;False;0.08454397720001908
-ek1_truncated;4;8;True;0.00011628279999058578
-ek1_truncated;8;8;False;0.028154431999973896
-ek1_truncated;8;8;True;0.0001611010000033275
-ek1_truncated;16;8;False;0.035056474199973306
-ek1_truncated;16;8;True;0.0003051037000204815
-ek1_truncated;32;8;False;0.049792988299986976
-ek1_truncated;32;8;True;0.0008990326000002824
-ek1_truncated;64;8;False;0.08139495049999823
-ek1_truncated;64;8;True;0.0048689479000131545
-ek1_truncated;128;8;False;0.17768519430001106
-ek1_truncated;128;8;True;0.027775210200024956
-ek0_reference;4;8;False;0.006150029100035681
-ek0_reference;4;8;True;0.0001585906999935105
-ek0_reference;8;8;False;0.006754524899997704
-ek0_reference;8;8;True;0.0007224639000014577
-ek0_reference;16;8;False;0.024076572599960856
-ek0_reference;16;8;True;0.009858659600013198
-ek0_reference;32;8;False;0.05583854480000809
-ek0_reference;32;8;True;0.04278747939997629
-ek0_reference;64;8;False;0.1704630323999936
-ek0_reference;64;8;True;0.2104338774000098
-ek0_reference;128;8;False;0.937638516800007
-ek0_reference;128;8;True;0.9407234414999948
-ek0_kronecker;4;8;False;0.00992896259999725
-ek0_kronecker;4;8;True;3.797660001509939e-05
-ek0_kronecker;8;8;False;0.010024652600031913
-ek0_kronecker;8;8;True;3.590629999052908e-05
-ek0_kronecker;16;8;False;0.009994579399972281
-ek0_kronecker;16;8;True;3.7471299992830606e-05
-ek0_kronecker;32;8;False;0.009970197899974664
-ek0_kronecker;32;8;True;3.796150003836374e-05
-ek0_kronecker;64;8;False;0.010039296100012507
-ek0_kronecker;64;8;True;4.381300000204647e-05
-ek0_kronecker;128;8;False;0.010083501700000852
-ek0_kronecker;128;8;True;5.382619997362781e-05
-ek0_kronecker;256;8;False;0.010101850099999865
-ek0_kronecker;256;8;True;8.703309999873454e-05
-ek0_kronecker;512;8;False;0.010689659900026526
-ek0_kronecker;512;8;True;0.000152005600011762
-ek0_kronecker;1024;8;False;0.012589014600007432
-ek0_kronecker;1024;8;True;0.0005045082999913575
