--- experiments/results/2_medium_scale_problem/hparams.json
@@ -1,17 +1,0 @@
-{
-  "dt":{
-    "0":0.5
-  },
-  "t0":{
-    "0":0.0
-  },
-  "tmax":{
-    "0":3.0
-  },
-  "forcing":{
-    "0":5.0
-  },
-  "ode_dimension":{
-    "0":10
-  }
-}
