--- experiments/results/2_medium_scale_problem/results.csv
@@ -1,31 +1,0 @@
-method;tolerance;nu;time_solve;deviation
-ek1_reference;1e-06;5;0.28051181499999966;1.4634241337693845e-10
-ek1_reference;1e-05;5;0.25563056100000026;1.31493707258835e-09
-ek1_reference;0.0001;5;0.2215188690000005;7.474116406687595e-09
-ek1_reference;0.001;5;0.18886839199999983;8.861735888910366e-08
-ek1_reference;0.01;5;0.16141254399999916;3.6774495768166283e-06
-ek1_reference;0.1;5;0.13576456000000015;6.293610619886967e-05
-ek1_diagonal;1e-06;5;0.5494651050000012;1.4637645513993305e-10
-ek1_diagonal;1e-05;5;0.49913895199999914;1.315506814643135e-09
-ek1_diagonal;0.0001;5;0.4996746590000001;7.47414526465394e-09
-ek1_diagonal;0.001;5;0.5386098620000013;8.861455385596859e-08
-ek1_diagonal;0.01;5;0.4343553480000004;3.6775029693904555e-06
-ek1_diagonal;0.1;5;0.42075075600000034;6.293652308063909e-05
-ek1_early_truncation;1e-06;5;1.2465939329999998;1.463755199295416e-10
-ek1_early_truncation;1e-05;5;0.9619025620000023;1.3154729795021471e-09
-ek1_early_truncation;0.0001;5;0.7034788540000001;7.474146695651075e-09
-ek1_early_truncation;0.001;5;0.563928446000002;8.861462154110832e-08
-ek1_early_truncation;0.01;5;0.39996725099999964;3.6775013545924885e-06
-ek1_early_truncation;0.1;5;0.20262079100000108;6.29365089574992e-05
-ek0_reference;1e-06;5;1.002771763999995;3.1389486987121053e-10
-ek0_reference;1e-05;5;0.9269483119999933;3.4389357620698316e-09
-ek0_reference;0.0001;5;0.9582856050000004;4.110763981654331e-08
-ek0_reference;0.001;5;0.7867714869999958;7.365575100463611e-07
-ek0_reference;0.01;5;0.6434194429999991;1.1266177158275456e-05
-ek0_reference;0.1;5;0.11603379399999625;0.0001939261903298738
-ek0_kronecker;1e-06;5;1.3531127019999971;3.148725047709906e-10
-ek0_kronecker;1e-05;5;1.3292831219999997;3.4393285069911683e-09
-ek0_kronecker;0.0001;5;1.3783827130000006;4.110774887152996e-08
-ek0_kronecker;0.001;5;1.1419729489999995;7.365402493246566e-07
-ek0_kronecker;0.01;5;0.9618742460000007;1.1265378986056807e-05
-ek0_kronecker;0.1;5;0.14698801600000877;0.00019392111217358545
