+++ plotting_only/1_sterilised_problem.py
