+++ setup.cfg
@@ -1,11 +1,13 @@
 [metadata]
-name = high-dimensional-ode-solvers
+name = hose
 version = 0.0.1
 author = Nicholas Krämer, Jonathan Schmidt, Nathanael Bosch
 license = MIT
 license_file = LICENSE.txt
 
 [options]
+package_dir =
+    =src
 packages = find:
 install_requires =
     importlib_metadata;python_version<"3.8"
