+++ src/hose/lines_and_ticks.mplstyle
